+++ packages/es-client/search.js
@@ -8,13 +8,13 @@
 
 const has = require('lodash/has');
 const omit = require('lodash/omit');
-const aws = require('aws-sdk');
-const { AmazonConnection } = require('aws-elasticsearch-connector');
+const { fromNodeProviderChain } = require('@aws-sdk/credential-providers');
 const elasticsearch = require('@elastic/elasticsearch');
 
 const { inTestMode } = require('@cumulus/common/test-utils');
 const Logger = require('@cumulus/logger');
 
+const createEsAmazonConnection = require('./esAmazonConnection');
 const queries = require('./queries');
 const aggs = require('./aggregations');
 
@@ -26,14 +26,7 @@ const logDetails = {
 const defaultIndexAlias = 'cumulus-alias';
 const multipleRecordFoundString = 'More than one record was found!';
 const recordNotFoundString = 'Record not found';
-
 const logger = new Logger({ sender: '@cumulus/es-client/search' });
-
-const getCredentials = () =>
-  new Promise((resolve, reject) => aws.config.getCredentials((err) => {
-    if (err) return reject(err);
-    return resolve();
-  }));
 
 /**
  * returns the local address of elasticsearch based on
@@ -58,8 +51,6 @@ const esTestConfig = () => ({
 });
 
 const esProdConfig = async (host) => {
-  if (!aws.config.credentials) await getCredentials();
-
   let node = 'http://localhost:9200';
 
   if (process.env.ES_HOST) {
@@ -68,12 +59,19 @@ const esProdConfig = async (host) => {
     node = `https://${host}`;
   }
 
+  const credentialsProvider = fromNodeProviderChain({
+    clientConfig: {
+      region: process.env.AWS_REGION,
+    },
+  });
+
+  const credentials = await credentialsProvider();
   return {
     node,
-    Connection: AmazonConnection,
-    awsConfig: {
-      credentials: aws.config.credentials,
-    },
+    ...createEsAmazonConnection({
+      credentials,
+      region: process.env.AWS_REGION,
+    }),
 
     // Note that this doesn't abort the query.
     requestTimeout: 50000, // milliseconds